# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Customer CRUD API",
    "version": "1.0.0",
    "description": "CRUD API over a customer dataset (customers). Data is persisted to a local JSON file and seeded from seed/<entity>.json on first read."
  },
  "servers": [{ "url": "/", "description": "Current server" }],
  "paths": {
    "/": {
      "get": {
        "summary": "List all entities with record counts",
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/{entity}": {
      "get": {
        "summary": "List all records in an entity",
        "parameters": [
          {
            "name": "entity",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "example": "customers" }
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "Unknown entity" }
        }
      },
      "post": {
        "summary": "Create a record",
        "parameters": [
          {
            "name": "entity",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "example": "customers" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Customer" }
            }
          }
        },
        "responses": {
          "201": { "description": "Created" },
          "400": { "description": "Missing id field" },
          "409": { "description": "Record already exists" }
        }
      }
    },
    "/{entity}/{id}": {
      "get": {
        "summary": "Get one record by id",
        "parameters": [
          {
            "name": "entity",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "example": "customers" }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "example": "CUST-100001" }
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "Record not found" }
        }
      },
      "put": {
        "summary": "Update a record (shallow merge)",
        "parameters": [
          {
            "name": "entity",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "example": "customers" }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "example": "CUST-100001" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object" } } }
        },
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "Record not found" }
        }
      },
      "delete": {
        "summary": "Delete a record",
        "parameters": [
          {
            "name": "entity",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "example": "customers" }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "example": "CUST-100001" }
          }
        ],
        "responses": {
          "204": { "description": "Deleted" },
          "404": { "description": "Record not found" }
        }
      }
    },
    "/api/GetOrdersForPhoneAIByPhoneNo": {
      "post": {
        "summary": "Look up orders by phone number (mirrors the real Jomashop endpoint)",
        "security": [{ "basicAuth": [] }],
        "parameters": [{ "name": "phoneno", "in": "query", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK" }, "404": { "description": "No customer found" } }
      }
    },
    "/api/GetOrdersForEmailAI": {
      "post": {
        "summary": "Look up orders by email (invented — no real endpoint shared yet)",
        "security": [{ "basicAuth": [] }],
        "parameters": [{ "name": "email", "in": "query", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK" }, "404": { "description": "No customer found" } }
      }
    },
    "/api/GetOrdersForNameAI": {
      "post": {
        "summary": "Look up orders by name (invented — no real endpoint shared yet)",
        "security": [{ "basicAuth": [] }],
        "parameters": [{ "name": "name", "in": "query", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK" }, "404": { "description": "No customer found" } }
      }
    },
    "/api/GetOrderDetailForPhoneAIByOrderNo": {
      "post": {
        "summary": "Get full order detail by order number (mirrors the real Jomashop endpoint)",
        "security": [{ "basicAuth": [] }],
        "parameters": [{ "name": "orderno", "in": "query", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK" }, "404": { "description": "Order not found" } }
      }
    },
    "/api/CancelOrderForPhoneAI": {
      "post": {
        "summary": "Cancel an order unconditionally (invented — CRM does not enforce SOP thresholds)",
        "security": [{ "basicAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["orderno"], "properties": { "orderno": { "type": "string" } } } } } },
        "responses": { "200": { "description": "OK" }, "404": { "description": "Order not found" } }
      }
    },
    "/api/RaiseZendeskTicket": {
      "post": {
        "summary": "Raise a Zendesk ticket (invented). orderno is optional — identity-verification-failed and out-of-scope tickets legitimately have no order to attach to.",
        "security": [{ "basicAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["type"],
                "properties": {
                  "orderno": { "type": "string" },
                  "type": { "type": "string" },
                  "reason": { "type": "string" },
                  "raised_by": { "type": "string" },
                  "exception": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Created" }, "404": { "description": "Order not found" } }
      }
    },
    "/api/CreateRMAForOrderAI": {
      "post": {
        "summary": "Create an RMA/return record (invented)",
        "security": [{ "basicAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["orderno"], "properties": { "orderno": { "type": "string" }, "reason": { "type": "string" } } } } } },
        "responses": { "201": { "description": "Created" }, "404": { "description": "Order not found" } }
      }
    },
    "/api/GetRMAHistoryByOrderNo": {
      "post": {
        "summary": "Get RMA count/history for an order (invented)",
        "security": [{ "basicAuth": [] }],
        "parameters": [{ "name": "orderno", "in": "query", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/GetTicketsByOrderNo": {
      "post": {
        "summary": "Get tickets raised for an order (invented)",
        "security": [{ "basicAuth": [] }],
        "parameters": [{ "name": "orderno", "in": "query", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/SendExtendClaimEmailAI": {
      "post": {
        "summary": "Record that an Extend claim email was sent (invented — not a Zendesk ticket)",
        "security": [{ "basicAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["orderno"], "properties": { "orderno": { "type": "string" }, "reason": { "type": "string" } } } } } },
        "responses": { "201": { "description": "Created" }, "404": { "description": "Order not found" } }
      }
    },
    "/api/AddNurixAICallLog": {
      "post": {
        "summary": "Log a call transcript (mirrors the real Jomashop endpoint)",
        "security": [{ "basicAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["Transcript", "Solved"],
                "properties": {
                  "PhoneNo": { "type": "string" },
                  "OrderNo": { "type": "string" },
                  "ItemNo": { "type": "string" },
                  "Email": { "type": "string" },
                  "Transcript": { "type": "string" },
                  "Solved": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Created" } }
      }
    },
    "/tools/calendar/add-business-days": {
      "post": {
        "summary": "Add N business days to a date (weekends skipped, no auth)",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["start_date", "business_days"], "properties": { "start_date": { "type": "string" }, "business_days": { "type": "number" } } } } } },
        "responses": { "200": { "description": "OK" }, "400": { "description": "Invalid input" } }
      }
    },
    "/tools/calendar/business-days-between": {
      "post": {
        "summary": "Count business days between two dates (weekends skipped, no auth)",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["date_a", "date_b"], "properties": { "date_a": { "type": "string" }, "date_b": { "type": "string" } } } } } },
        "responses": { "200": { "description": "OK" }, "400": { "description": "Invalid input" } }
      }
    }
  },
  "components": {
    "schemas": {
      "Customer": {
        "type": "object",
        "required": ["customer_id"],
        "properties": {
          "customer_id": { "type": "string" },
          "profile": { "type": "object" },
          "contact": { "type": "object" },
          "addresses": { "type": "object" },
          "verification": { "type": "object" },
          "account": { "type": "object" },
          "preferences": { "type": "object" },
          "summary": { "type": "object" },
          "metadata": { "type": "object" }
        }
      }
    },
    "securitySchemes": {
      "basicAuth": { "type": "http", "scheme": "basic" }
    }
  }
}